@@ -15,7 +15,7 @@ import org.springframework.security.core.Authentication;
 import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Component;
 
-import javax.transaction.TransactionalException;
+import jakarta.transaction.TransactionalException;
 import java.util.Arrays;
 import java.util.Random;
 
